# WikidataGame/WikidataGame.ApiClient/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Wikidata Game API"},"paths":{"/api/Auth":{"get":{"tags":["Auth"],"summary":"Authenticates a player using username/password","operationId":"Authenticate","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"X-Username","in":"header","description":"Username (min. 3 chars)","required":false,"type":"string"},{"name":"X-Password","in":"header","description":"Password (min. 8 chars)","required":false,"type":"string"},{"name":"X-Push-Token","in":"header","description":"push token generated through firebase/apns","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/AuthInfo"}}}}},"/api/Auth/Google":{"get":{"tags":["Auth"],"summary":"Authenticates a player using Google Play Services","operationId":"AuthenticateGooglePlay","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"X-Username","in":"header","description":"Username (min. 3 chars)","required":false,"type":"string"},{"name":"X-Auth-Code","in":"header","description":"backend auth code","required":false,"type":"string"},{"name":"X-Push-Token","in":"header","description":"push token generated through firebase/apns","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/AuthInfo"}}}}},"/api/Friends":{"get":{"tags":["Friends"],"summary":"Retrieves the friendlist for the signed in user","operationId":"GetFriends","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Player"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"post":{"tags":["Friends"],"summary":"Adds a user as a friend","operationId":"PostFriend","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"friendId","in":"query","description":"The user id of the player that should be added","required":false,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/Player"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Friends/{friendId}":{"delete":{"tags":["Friends"],"summary":"Removes a friend from the friend list","operationId":"DeleteFriend","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"friendId","in":"path","description":"User id of the friend to be removed","required":true,"type":"string"}],"responses":{"204":{"description":"Success","schema":{"$ref":"#/definitions/Player"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Friends/Find":{"get":{"tags":["Friends"],"summary":"Retrieves users (limit 10) with a username similar to the query string","operationId":"GetFindFriends","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"query","in":"query","description":"Username or part of a username (min. 3 characters)","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Player"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/GameRequests":{"get":{"tags":["GameRequests"],"summary":"Retrieves all game requests for the authenticated player","operationId":"GetGameRequests","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GameRequestList"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"post":{"tags":["GameRequests"],"summary":"Sends a game request to the specified user","operationId":"RequestMatch","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"userId","in":"query","description":"user id","required":false,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/GameRequest"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"delete":{"tags":["GameRequests"],"summary":"Deletes a game request for the sender or recipient","operationId":"DeleteGameRequest","consumes":[],"produces":[],"parameters":[{"name":"gameRequestId","in":"query","description":"game request identifier","required":false,"type":"string"}],"responses":{"204":{"description":"Success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Games/AcceptRequest":{"post":{"tags":["Games"],"summary":"Creates a new game by accepting a game request","operationId":"CreateNewGameByRequest","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"gameRequestId","in":"query","description":"game request identifier","required":false,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/GameInfo"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Games":{"get":{"tags":["Games"],"summary":"Retrieves all currently running games for the authenticated player","operationId":"GetGames","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/GameInfo"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"post":{"tags":["Games"],"summary":"Creates a new game and matches the player with an opponent","operationId":"CreateNewGame","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/GameInfo"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Games/{gameId}":{"get":{"tags":["Games"],"summary":"Retrieves information on the specified game","operationId":"RetrieveGameState","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"gameId","in":"path","description":"game identifier","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Game"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"delete":{"tags":["Games"],"summary":"Stops/deletes the specified game","operationId":"DeleteGame","consumes":[],"produces":[],"parameters":[{"name":"gameId","in":"path","description":"game identifier","required":true,"type":"string"}],"responses":{"204":{"description":"Success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/games/{gameId}/minigames":{"post":{"tags":["Minigames"],"summary":"Initializes a new minigame","operationId":"InitalizeMinigame","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"gameId","in":"path","description":"game identifier","required":true,"type":"string"},{"name":"minigameParams","in":"body","description":"minigame information containing category and tile identifier","required":false,"schema":{"$ref":"#/definitions/MiniGameInit"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/MiniGame"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/games/{gameId}/minigames/{minigameId}":{"get":{"tags":["Minigames"],"summary":"Retrieves the details of the specified minigame","operationId":"RetrieveMinigameInfo","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"gameId","in":"path","description":"game identifier","required":true,"type":"string"},{"name":"minigameId","in":"path","description":"minigame identifier","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/MiniGame"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]},"post":{"tags":["Minigames"],"summary":"Post the answer(s) to the specified minigame","operationId":"AnswerMinigame","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"gameId","in":"path","description":"game identifier","required":true,"type":"string"},{"name":"minigameId","in":"path","description":"minigame identifier","required":true,"type":"string"},{"name":"answers","in":"body","description":"answer(s)","required":false,"schema":{"uniqueItems":false,"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/MiniGameResult"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2":[]}]}},"/api/Platform/Stats":{"get":{"tags":["Platform"],"summary":"Retrieves statistics containing the number of categories, games played and questions added","operationId":"GetPlatformStats","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PlatformStats"}}}}},"/api/Platform/Minigame/{minigameId}":{"get":{"tags":["Platform"],"summary":"Retrieves detailed information on a bygone minigame by id","operationId":"GetPlatformMinigameById","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"minigameId","in":"path","description":"minigame identifier","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/DetailedMiniGame"}}}}},"/api/Platform/Question":{"get":{"tags":["Platform"],"summary":"Retrieves a list of all available questions","operationId":"GetPlatformQuestions","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Question"}}}}},"post":{"tags":["Platform"],"summary":"Adds the specified question to the question catalogue (with status pending)","operationId":"AddPlatformQuestion","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"question","in":"body","description":"question to add","required":false,"schema":{"$ref":"#/definitions/Question"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/Question"}}}}},"/api/Platform/Question/{questionId}/Rating":{"post":{"tags":["Platform"],"summary":"Adds a rating for the specified question","operationId":"AddPlatformQuestionRating","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"questionId","in":"path","description":"question identifier","required":true,"type":"string"},{"name":"rating","in":"query","description":"rating (1-5)","required":true,"type":"integer","format":"int32","maximum":5.0,"minimum":1.0}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Question"}}}}},"/api/Platform/Category":{"get":{"tags":["Platform"],"summary":"Retrieves a list of all available categories","operationId":"GetPlatformCategories","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Category"}}}}}}},"definitions":{"AuthInfo":{"type":"object","properties":{"bearer":{"type":"string"},"expires":{"format":"date-time","type":"string"},"user":{"$ref":"#/definitions/Player"}}},"Player":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"profileImage":{"type":"string"}}},"GameRequestList":{"type":"object","properties":{"incoming":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/GameRequest"}},"outgoing":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/GameRequest"}}}},"GameRequest":{"type":"object","properties":{"id":{"type":"string"},"sender":{"$ref":"#/definitions/Player"},"recipient":{"$ref":"#/definitions/Player"}}},"GameInfo":{"type":"object","properties":{"gameId":{"type":"string"},"isAwaitingOpponentToJoin":{"type":"boolean"},"message":{"type":"string"}}},"Game":{"type":"object","properties":{"id":{"type":"string"},"tiles":{"uniqueItems":false,"type":"array","items":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Tile"}}},"winningPlayerIds":{"uniqueItems":false,"type":"array","items":{"type":"string"}},"me":{"$ref":"#/definitions/Player"},"opponent":{"$ref":"#/definitions/Player"},"nextMovePlayerId":{"type":"string"},"moveExpiry":{"format":"date-time","type":"string"},"awaitingOpponentToJoin":{"type":"boolean"}}},"Tile":{"type":"object","properties":{"id":{"type":"string"},"chosenCategoryId":{"type":"string"},"availableCategories":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Category"}},"difficulty":{"format":"int32","type":"integer"},"ownerId":{"type":"string"}}},"Category":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"}}},"MiniGameInit":{"type":"object","properties":{"tileId":{"type":"string"},"categoryId":{"type":"string"}}},"MiniGame":{"type":"object","properties":{"id":{"type":"string"},"type":{"format":"int32","enum":[0,1,2],"type":"integer"},"taskDescription":{"type":"string"},"answerOptions":{"uniqueItems":false,"type":"array","items":{"type":"string"}}}},"MiniGameResult":{"type":"object","properties":{"isWin":{"type":"boolean"},"correctAnswer":{"uniqueItems":false,"type":"array","items":{"type":"string"}},"tiles":{"uniqueItems":false,"type":"array","items":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Tile"}}},"nextMovePlayerId":{"type":"string"}}},"PlatformStats":{"type":"object","properties":{"numberOfQuestions":{"format":"int32","type":"integer"},"numberOfCategories":{"format":"int32","type":"integer"},"numberOfGamesPlayed":{"format":"int32","type":"integer"},"numberOfContributions":{"format":"int32","type":"integer"}}},"DetailedMiniGame":{"type":"object","properties":{"question":{"$ref":"#/definitions/Question"},"correctAnswer":{"uniqueItems":false,"type":"array","items":{"type":"string"}},"id":{"type":"string"},"type":{"format":"int32","enum":[0,1,2],"type":"integer"},"taskDescription":{"type":"string"},"answerOptions":{"uniqueItems":false,"type":"array","items":{"type":"string"}}}},"Question":{"required":["sparqlQuery","taskDescription","category","miniGameType"],"type":"object","properties":{"id":{"type":"string"},"sparqlQuery":{"type":"string"},"taskDescription":{"type":"string"},"category":{"$ref":"#/definitions/Category"},"miniGameType":{"format":"int32","enum":[0,1,2],"type":"integer"},"status":{"format":"int32","enum":[0,1,2],"type":"integer"},"rating":{"format":"double","type":"number"}}}},"securityDefinitions":{"oauth2":{"name":"Authorization","in":"header","type":"apiKey","description":"JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""}}}